+++ scripts/init/tf/prepare/yc.py
@@ -1,0 +1,2 @@
+def prepare_yc() -> bool:
+    return False